refactor cube and precisions
