+++ benchmark/src/main/resources/benchmark_policy.json
@@ -30,7 +30,7 @@
   ],
   "cubes": [
     {
-      "precisions": [
+      "dimensions": [
         {
           "dimensionName": "time",
           "bucketType": "second"
