+++ examples/policies/IKafkaDirect-OPrint.json
@@ -27,8 +27,8 @@
   ],
   "cubes": [
     {
-      "cube": "testCube",
-      "precisions": [
+      "name": "testCube",
+      "dimensions": [
         {
           "dimension": "message"
         }
