+++ examples/policies/ITwitter-OParquet.json
@@ -39,8 +39,8 @@
   ],
   "cubes": [
     {
-      "cube": "twitter",
-      "precisions": [
+      "name": "twitter",
+      "dimensions": [
         {
           "dimension": "status",
           "precision": "hastags"
