+++ examples/policies/ITwitter-OPrintWithFragments.json
@@ -48,8 +48,8 @@
   ],
   "cubes": [
     {
-      "cube": "twitter",
-      "precisions": [
+      "name": "twitter",
+      "dimensions": [
         {
           "dimension": "status",
           "precision": "hastags"
