+++ examples/policies/ITwitter-ORedis.json
@@ -39,8 +39,8 @@
   ],
   "cubes": [
     {
-      "cube": "twitter",
-      "precisions": [
+      "name": "twitter",
+      "dimensions": [
         {
           "dimension": "status",
           "precision": "hastags"
