+++ plugins/dimension-default/pom.xml
@@ -26,7 +26,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>dimension-native</artifactId>
+    <artifactId>dimension-default</artifactId>
 
 
 </project>
